// src/app/shared/model/profile/profile.response.ts
export class ProfileResponse {
  id: number;
  userName: string;
  password: string;
  email: string;
  firstName: string;
  lastName: string;
  phone: string;
  address: string;
  image: string;
  role: string;
  createdDate: string;
  sex: string;
  enabled: boolean;
  status: boolean;
  dateOfBirth: string;
  id_creator: number;
  id_homeStay: number;
  homeStayName: string;
  }